Make CommandProcessorMeta supproting async
