+++ examples/this-that/app/src/lib.rs
@@ -5,7 +5,7 @@ use gmeta::{InOut, Metadata};
 use gstd::debug;
 use gstd::prelude::*;
 use sails_macros::{command_handlers, query_handlers};
-use sails_service::SimpleService;
+use sails_service::{BoxedFuture, SimpleService};
 
 pub struct ProgramMetadata;
 
@@ -22,21 +22,18 @@ impl Metadata for ProgramMetadata {
 // TODO: The 2 below structs and their impls could be generated by the `commmand_handlers!`/`query_handlers!` macros
 pub struct CommandProcessorMeta;
 
-impl sails_service::RequestProcessorMeta for CommandProcessorMeta {
+impl sails_service::CommandProcessorMeta for CommandProcessorMeta {
     type Request = commands::Commands;
     type Response = commands::CommandResponses;
-    type ProcessFn = fn(Self::Request) -> (Self::Response, bool);
-    // For async processing
-    //type Handle = fn(Self::Requests) -> BoxedFuture<Self::Responses>;
+    type ProcessFn = fn(Self::Request) -> BoxedFuture<(Self::Response, bool)>;
 }
 
 pub struct QueryProcessorMeta;
 
-impl sails_service::RequestProcessorMeta for QueryProcessorMeta {
+impl sails_service::QueryProcessorMeta for QueryProcessorMeta {
     type Request = queries::Queries;
     type Response = queries::QueryResponses;
     type ProcessFn = fn(Self::Request) -> (Self::Response, bool);
-    //type Handle = fn(Self::Requests) -> BoxedFuture<Self::Responses>;
 }
 
 pub type Service = SimpleService<CommandProcessorMeta, QueryProcessorMeta>;
